+++ sql/queries/feed_follows.sql
@@ -11,6 +11,7 @@ SELECT
 FROM inserted_feed_follow
 INNER JOIN feeds ON inserted_feed_follow.feed_id = feeds.id
 INNER JOIN users ON inserted_feed_follow.user_id = users.id;
+--
 
 -- name: GetFeedFollowsForUser :many
 SELECT feed_follows.*, feeds.name AS feed_name, users.name AS user_name
@@ -18,6 +19,8 @@ FROM feed_follows
 INNER JOIN feeds ON feed_follows.feed_id = feeds.id
 INNER JOIN users ON feeds.user_id = users.id
 WHERE feed_follows.user_id = $1;
+--
 
--- name: DeleteFeedFollows :exec
-DELETE FROM feed_follows;
+-- name: DeleteFeedFollow :exec
+DELETE FROM feed_follows WHERE feed_id = $1 AND user_id = $2;
+--
